+++ packages/terra-core-docs/src/terra-dev-site/doc/form-select/example/search/Search.jsx
@@ -8,7 +8,7 @@ const cx = classNames.bind(styles);
 const SearchSelectExample = () => (
   <>
     <label htmlFor="color-field-1"><strong>Colors</strong></label>
-    <SearchSelect placeholder="Select a color" className={cx('form-select')} inputId="color-field-1">
+    <SearchSelect placeholder="Select a color" ariaLabel="Select a color" className={cx('form-select')} inputId="color-field-1">
       <SearchSelect.Option value="blue" display="Blue" />
       <SearchSelect.Option value="green" display="Green" />
       <SearchSelect.Option value="purple" display="Purple" />